Move matrix summarization Ops to sub-module
